+++ lib/system_idle.dart
@@ -1,3 +1,35 @@
-library system_idle;
+import "dart:io";
 
-export 'src/system_idle.dart';
+import "src/interface.dart";
+import "src/system_idle_linux.dart";
+import "src/system_idle_macos.dart";
+import "src/system_idle_windows.dart";
+
+class SystemIdle {
+  static const defaultIdleDuration = Duration(minutes: 5);
+
+  SystemIdle._() {
+    if (Platform.isWindows) {
+      _systemIdleChecker = SystemIdleCheckerWindows();
+    } else if (Platform.isMacOS) {
+      _systemIdleChecker = SystemIdleCheckerMacOS();
+    } else if (Platform.isLinux) {
+      _systemIdleChecker = SystemIdleCheckerLinux();
+    }
+  }
+
+  static final SystemIdle _instance = SystemIdle._();
+
+  /// A singleton instance of [SystemIdle] class
+  static SystemIdle get instance => _instance;
+
+  late SystemIdleChecker _systemIdleChecker;
+
+  /// {@macro system_idle_initialize}
+  Future<void> initialize({Duration duration = defaultIdleDuration}) =>
+      _systemIdleChecker.initialize(duration: duration);
+
+  /// {@macro system_idle_on_idle_state_changed}
+  Stream<bool> get onIdleStateChanged =>
+      _systemIdleChecker.onIdleStateChanged();
+}