+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+* Added `Future<Duration> getIdleDuration()`
+
+## 1.0.0
+* Change `initialize` to take a `Duration` instead of integer seconds
+
 ## 0.2.1
 * Bump win32 version to 3.1.3